+++ examples/hello/restful-hello-world.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"io"
+	"log"
+	"net/http"
+
+	restful "github.com/emicklei/go-restful/v3"
+)
+
+// This example shows the minimal code needed to get a restful.WebService working.
+//
+// GET http://localhost:8080/hello
+
+func main() {
+	ws := new(restful.WebService)
+	ws.Route(ws.GET("/hello").To(hello))
+	restful.Add(ws)
+	log.Fatal(http.ListenAndServe(":8080", nil))
+}
+
+func hello(req *restful.Request, resp *restful.Response) {
+	io.WriteString(resp, "world")
+}